+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Add `transaction plt mint` command to mint protocol level tokens by the governance account.
+- Add `transaction plt burn` command to burn protocol level tokens by the governance account.
+- Refactore plt commands into its own subcommand.
+
 ## 9.0.0 - 2025-05-09
 
 - Add `raw GetTokenInfo` command to display the info of a protocol level token.